[#946] upload apk improvements
* added build.gradle
* removed libs
* added more info to deploy
* fixed deploy script
