--- util/upload-apk/.gitignore
@@ -1,1 +1,0 @@
-/bin
